test: add tests for retry_download function This adds two different test cases for retry_download to verify: - It succeeds after failed attempts - It fails/retries for a maximum of 5 times

+++ tests/download.bats
@@ -4,13 +4,18 @@ load "${BATS_PLUGIN_PATH}/load.bash"
 
 load '../lib/download.bash'
 
+is_stubbed=false
+sleep_stub_command='echo sleep $@'
+sleep_stubs=("$sleep_stub_command" "$sleep_stub_command" "$sleep_stub_command" "$sleep_stub_command" "$sleep_stub_command")
+
+
 #
 # Tests for top-level docker bootstrap command. The rest of the plugin runs in Go.
 #
 
 # Uncomment the following line to debug stub failures
 # export [stub_command]_STUB_DEBUG=/dev/tty
-#export DOCKER_STUB_DEBUG=/dev/tty
+#export SLEEP_STUB_DEBUG=/dev/tty
 
 setup() {
   export BUILDKITE_PLUGIN_ECR_SCAN_RESULTS_BUILDKITE_PLUGIN_TEST_MODE=true
@@ -18,7 +23,12 @@ setup() {
 
 teardown() {
     unset BUILDKITE_PLUGIN_ECR_SCAN_RESULTS_BUILDKITE_PLUGIN_TEST_MODE
-    rm ./ecr-scan-results-buildkite-plugin || true
+    if $is_stubbed; then
+      unstub curl || true
+      unstub sleep || true
+    else
+     rm ./ecr-scan-results-buildkite-plugin || true
+    fi
 }
 
 create_script() {
@@ -45,4 +55,50 @@ EOM
   assert_success
   assert_line --regexp "https://github.com/cultureamp/ecr-scan-results-buildkite-plugin/releases/latest/download/ecr-scan-results-buildkite-plugin_linux_amd64 ecr-scan-results-buildkite-plugin"
   assert_line --regexp "executing ecr-scan-results-buildkite-plugin"
+}
+
+@test "Attempts to download and succeed after 2 tries" {
+
+  # Fails for first 2 tries and passes on later 
+
+  stub curl \
+      "echo 'curl 1 - Could not resolve host' && exit 6" \
+      "echo 'curl 2 - Failed to connect to example.com port 80 - Connection refused' && exit 7" \
+      "echo 'curl 3 - Success' && exit 0"
+
+  stub sleep "${sleep_stubs[@]}"
+  is_stubbed=true
+
+  # executes the retry download command
+  run retry_download "https://example.com" "$TMPDIR/output-plugin" "curl"
+
+  
+  assert_success
+  assert_line --partial "Attempt 1"
+  assert_line --partial "Attempt 2"
+  assert_line --partial "Success"
+
+}
+
+@test "Attempts to download and fails after maximum tries" {
+
+  # Fails for all 5 (maximum) tries 
+
+  stub curl \
+        "echo 'curl 1 - Could not resolve host: example.com' && exit 6" \
+        "echo 'curl 2 (7) Failed to connect to example.com port 80 - Connection refused' && exit 7" \
+        "echo 'curl 3 (7) Failed to connect to example.com port 80 - Connection refused' && exit 7"  \
+        "echo 'curl 4 (7) Failed to connect to example.com port 80 - Connection refused' && exit 7"  \
+        "echo 'curl 5 (7) Failed to connect to example.com port 80 - Connection refused' && exit 7"  \
+        "echo 'curl 6 (7) Failed to connect to example.com port 80 - Connection refused' && exit 7"  \
+
+  stub sleep "${sleep_stubs[@]}"
+  is_stubbed=true
+
+  # executes the retry download command
+  run retry_download "https://example.com" "$TMPDIR/output-plugin" "curl"
+
+  assert_failure 1
+  #Checks if multiple attempts are made to retry the download
+  assert_line --partial "Download failed after 5 attempts" 
 }
